fix: Remove deprecated call to `get_cached_token`
The `spotipy` library has deprecated the direct use of `get_cached_token`
on the `SpotifyOAuth` object. This commit replaces the deprecated call
with the recommended approach of using `validate_token` in combination
with `cache_handler.get_cached_token`.

This resolves the `DeprecationWarning` that was showing up when running
the tests.